fix camera shared timer windows cleanups
git-svn-id: svn://84.194.112.164/bengine@333 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/utils/events.cpp
@@ -10,7 +10,7 @@ Events::Events()
 {
 }
 
-sharedTimer* Events::registerSharedtimer(float responsetime)
+sharedTimer* Events::registerSharedtimer(unsigned int responsetime)
 {
 	// create a new event and a pointer to it
 	sharedtimers.push_back(sharedTimer());
